requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
pdfplumber
python-docx
pydantic-settings
python-multipart


